move gaestream into it's own profile as it can interfer with electron builds
