+++ libs/logging/src/streams/gaestream.ts
