+++ libs/logging/src/streams/seqstream.ts
